Rename responses namespace
