+++ src/Patpass/PatpassByWebpay/Responses/TransactionCreateResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Transbank\Patpass\PatpassByWebpay;
+namespace Transbank\Patpass\PatpassByWebpay\Responses;
 
 class TransactionCreateResponse
 {
