+++ src/Patpass/PatpassByWebpay/Responses/TransactionStatusResponse.php
@@ -1,0 +1,7 @@
+<?php
+namespace Transbank\Patpass\PatpassByWebpay\Responses;
+
+class TransactionStatusResponse extends \Transbank\Webpay\WebpayPlus\Responses\TransactionStatusResponse
+{
+
+}
