+++ src/Patpass/PatpassByWebpay/Transaction.php
@@ -5,6 +5,7 @@ use Transbank\Patpass\PatpassByWebpay;
 use Transbank\Patpass\PatpassByWebpay\Exceptions\TransactionCreateException;
 use Transbank\Patpass\PatpassByWebpay\Exceptions\TransactionCommitException;
 use Transbank\Patpass\PatpassByWebpay\Exceptions\TransactionStatusException;
+use Transbank\Patpass\PatpassByWebpay\Responses\TransactionCommitResponse;
 
 class Transaction
 {
