--- src/Patpass/PatpassByWebpay/TransactionStatusResponse.php
@@ -1,7 +1,0 @@
-<?php
-namespace Transbank\Patpass\PatpassByWebpay;
-
-class TransactionStatusResponse extends \Transbank\Webpay\WebpayPlus\TransactionStatusResponse
-{
-
-}
