+++ src/TransaccionCompleta/Responses/MallTransactionCommitResponse.php
@@ -1,15 +1,6 @@
 <?php
 
-/**
- * Class MallTransactionCommitResponse
- *
- * @category
- * @package Transbank\TransaccionCompleta
- *
- */
-
-
-namespace Transbank\TransaccionCompleta;
+namespace Transbank\TransaccionCompleta\Responses;
 
 use Transbank\Utils\Utils;
