+++ src/TransaccionCompleta/Responses/MallTransactionCreateResponse.php
@@ -1,28 +1,18 @@
 <?php
-
-/**
- * Class TransactionCreateResponse
- *
- * @category
- * @package Transbank\TransaccionCompleta
- *
- */
-
-
-namespace Transbank\TransaccionCompleta;
+namespace Transbank\TransaccionCompleta\Responses;
 
 use Transbank\Utils\Utils;
 
-class TransactionCreateResponse
+class MallTransactionCreateResponse
 {
     public $token;
-
+    
     public function __construct($json)
     {
-        $token =  Utils::returnValueIfExists($json, "token");
+        $token = Utils::returnValueIfExists($json, "token");
         $this->setToken($token);
     }
-
+    
     /**
      * @return mixed
      */
@@ -30,13 +20,15 @@ class TransactionCreateResponse
     {
         return $this->token;
     }
-
+    
     /**
      * @param mixed $token
+     * @return MallTransactionCreateResponse
      */
     public function setToken($token)
     {
         $this->token = $token;
+        
         return $this;
     }
 }
