+++ src/TransaccionCompleta/Responses/MallTransactionInstallmentsResponse.php
@@ -1,15 +1,6 @@
 <?php
 
-/**
- * Class MallTransactionInstallmentsResponse
- *
- * @category
- * @package Transbank\TransaccionCompleta
- *
- */
-
-
-namespace Transbank\TransaccionCompleta;
+namespace Transbank\TransaccionCompleta\Responses;
 
 use Transbank\Utils\Utils;
