+++ src/TransaccionCompleta/Responses/MallTransactionRefundResponse.php
@@ -1,19 +1,9 @@
 <?php
-
-/**
- * Class TransactionRefundResponse
- *
- * @category
- * @package Transbank\TransaccionCompleta
- *
- */
-
-
-namespace Transbank\TransaccionCompleta;
+namespace Transbank\TransaccionCompleta\Responses;
 
 use Transbank\Utils\Utils;
 
-class TransactionRefundResponse
+class MallTransactionRefundResponse
 {
     public $type;
     public $authorizationCode;
@@ -21,23 +11,23 @@ class TransactionRefundResponse
     public $nullifiedAmount;
     public $balance;
     public $responseCode;
-
+    
     public function __construct($json)
     {
-        $type =  Utils::returnValueIfExists($json, "type");
+        $type = Utils::returnValueIfExists($json, "type");
         $this->setType($type);
-        $authorizationCode =  Utils::returnValueIfExists($json, "authorization_code");
+        $authorizationCode = Utils::returnValueIfExists($json, "authorization_code");
         $this->setAuthorizationCode($authorizationCode);
-        $authorizationDate =  Utils::returnValueIfExists($json, "authorization_date");
+        $authorizationDate = Utils::returnValueIfExists($json, "authorization_date");
         $this->setAuthorizationDate($authorizationDate);
-        $nullifiedAmount =  Utils::returnValueIfExists($json, "nullified_amount");
+        $nullifiedAmount = Utils::returnValueIfExists($json, "nullified_amount");
         $this->setNullifiedAmount($nullifiedAmount);
-        $balance =  Utils::returnValueIfExists($json, "balance");
+        $balance = Utils::returnValueIfExists($json, "balance");
         $this->setBalance($balance);
-        $responseCode =  Utils::returnValueIfExists($json, "response_code");
+        $responseCode = Utils::returnValueIfExists($json, "response_code");
         $this->setResponseCode($responseCode);
     }
-
+    
     /**
      * @return mixed
      */
@@ -45,17 +35,18 @@ class TransactionRefundResponse
     {
         return $this->type;
     }
-
+    
     /**
      * @param mixed $type
-     * @return TransactionRefundResponse
+     * @return MallTransactionRefundResponse
      */
     public function setType($type)
     {
         $this->type = $type;
+        
         return $this;
     }
-
+    
     /**
      * @return mixed
      */
@@ -63,17 +54,18 @@ class TransactionRefundResponse
     {
         return $this->authorizationCode;
     }
-
+    
     /**
      * @param mixed $authorizationCode
-     * @return TransactionRefundResponse
+     * @return MallTransactionRefundResponse
      */
     public function setAuthorizationCode($authorizationCode)
     {
         $this->authorizationCode = $authorizationCode;
+        
         return $this;
     }
-
+    
     /**
      * @return mixed
      */
@@ -81,17 +73,18 @@ class TransactionRefundResponse
     {
         return $this->authorizationDate;
     }
-
+    
     /**
      * @param mixed $authorizationDate
-     * @return TransactionRefundResponse
+     * @return MallTransactionRefundResponse
      */
     public function setAuthorizationDate($authorizationDate)
     {
         $this->authorizationDate = $authorizationDate;
+        
         return $this;
     }
-
+    
     /**
      * @return mixed
      */
@@ -99,17 +92,18 @@ class TransactionRefundResponse
     {
         return $this->nullifiedAmount;
     }
-
+    
     /**
      * @param mixed $nullifiedAmount
-     * @return TransactionRefundResponse
+     * @return MallTransactionRefundResponse
      */
     public function setNullifiedAmount($nullifiedAmount)
     {
         $this->nullifiedAmount = $nullifiedAmount;
+        
         return $this;
     }
-
+    
     /**
      * @return mixed
      */
@@ -117,17 +111,18 @@ class TransactionRefundResponse
     {
         return $this->balance;
     }
-
+    
     /**
      * @param mixed $balance
-     * @return TransactionRefundResponse
+     * @return MallTransactionRefundResponse
      */
     public function setBalance($balance)
     {
         $this->balance = $balance;
+        
         return $this;
     }
-
+    
     /**
      * @return mixed
      */
@@ -135,16 +130,17 @@ class TransactionRefundResponse
     {
         return $this->responseCode;
     }
-
+    
     /**
      * @param mixed $responseCode
      *
      *
-     * @return TransactionRefundResponse
+     * @return MallTransactionRefundResponse
      */
     public function setResponseCode($responseCode)
     {
         $this->responseCode = $responseCode;
+        
         return $this;
     }
 }
