+++ src/TransaccionCompleta/Responses/TransactionInstallmentsResponse.php
@@ -1,15 +1,5 @@
 <?php
-
-/**
- * Class TransactionInstallmentsResponse
- *
- * @category
- * @package Transbank\TransaccionCompleta
- *
- */
-
-
-namespace Transbank\TransaccionCompleta;
+namespace Transbank\TransaccionCompleta\Responses;
 
 use Transbank\Utils\Utils;
