+++ src/Webpay/Modal/Responses/TransactionRefundResponse.php
@@ -2,7 +2,7 @@
 
 namespace Transbank\Webpay\Modal\Responses;
 
-class TransactionRefundResponse extends \Transbank\Webpay\WebpayPlus\TransactionRefundResponse
+class TransactionRefundResponse extends \Transbank\Webpay\WebpayPlus\Responses\TransactionRefundResponse
 {
 
 }
