+++ src/Webpay/WebpayPlus/Responses/TransactionCommitResponse.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Transbank\Webpay\WebpayPlus\Responses;
+
+use Transbank\Webpay\WebpayPlus\Responses\TransactionStatusResponse;
+
+class TransactionCommitResponse extends TransactionStatusResponse
+{
+
+}
