+++ src/Webpay/WebpayPlus/Responses/TransactionStatusResponse.php
@@ -1,17 +1,7 @@
 <?php
 
-/**
- * Class TransactionStatusResponse
- *
- * @category
- * @package Transbank\TransaccionCompleta
- *
- */
 
-
-namespace Transbank\TransaccionCompleta;
-
-use Transbank\Utils\Utils;
+namespace Transbank\Webpay\WebpayPlus\Responses;
 
 class TransactionStatusResponse
 {
@@ -20,44 +10,35 @@ class TransactionStatusResponse
     public $status;
     public $buyOrder;
     public $sessionId;
-    public $cardDetail; // card_number
+    public $cardNumber;
+    public $cardDetail;
     public $accountingDate;
     public $transactionDate;
     public $authorizationCode;
     public $paymentTypeCode;
     public $responseCode;
-    public $installmentsNumber;
     public $installmentsAmount;
+    public $installmentsNumber;
+    public $balance;
 
 
     public function __construct($json)
     {
-        $vci =  Utils::returnValueIfExists($json, "vci");
-        $this->setVci($vci);
-        $amount =  Utils::returnValueIfExists($json, "amount");
-        $this->setAmount($amount);
-        $status =  Utils::returnValueIfExists($json, "status");
-        $this->setStatus($status);
-        $buyOrder =  Utils::returnValueIfExists($json, "buy_order");
-        $this->setBuyOrder($buyOrder);
-        $sessionId = Utils::returnValueIfExists($json, "session_id");
-        $this->setSessionId($sessionId);
-        $cardDetail= Utils::returnValueIfExists($json, "card_detail");
-        $this->setCardDetail($cardDetail);
-        $accountingDate = Utils::returnValueIfExists($json, "accounting_date");
-        $this->setAccountingDate($accountingDate);
-        $transactionDate = Utils::returnValueIfExists($json, "transaction_date");
-        $this->setTransactionDate($transactionDate);
-        $authorizationCode = Utils::returnValueIfExists($json, "authorization_code");
-        $this->setAuthorizationCode($authorizationCode);
-        $paymentTypeCode = Utils::returnValueIfExists($json, "payment_type_code");
-        $this->setPaymentTypeCode($paymentTypeCode);
-        $responseCode = Utils::returnValueIfExists($json, "response_code");
-        $this->setResponseCode($responseCode);
-        $installmentsNumber = Utils::returnValueIfExists($json, "installments_number");
-        $this->setInstallmentsNumber($installmentsNumber);
-        $installmentsAmount = Utils::returnValueIfExists($json, "installments_amount");
-        $this->setInstallmentsAmount($installmentsAmount);
+        $this->vci = isset($json["vci"]) ? $json["vci"] : null;
+        $this->amount = isset($json["amount"]) ? $json["amount"] : null;
+        $this->status = isset($json["status"]) ? $json["status"] : null;
+        $this->buyOrder = isset($json["buy_order"]) ? $json["buy_order"] : null;
+        $this->sessionId = isset($json["session_id"]) ?$json["session_id"] : null;
+        $this->cardNumber = isset($json["card_detail"]) && isset($json["card_detail"]["card_number"]) ? $json["card_detail"]["card_number"] : null;
+        $this->cardDetail = isset($json["card_detail"]) ? $json["card_detail"] : null;
+        $this->accountingDate = isset($json["accounting_date"]) ? $json["accounting_date"] : null;
+        $this->transactionDate = isset($json["transaction_date"]) ? $json["transaction_date"] : null;
+        $this->authorizationCode = isset($json["authorization_code"]) ? $json["authorization_code"] : null;
+        $this->paymentTypeCode = isset($json["payment_type_code"]) ? $json["payment_type_code"] : null;
+        $this->responseCode = isset($json["response_code"]) ? $json["response_code"] : null;
+        $this->installmentsAmount = isset($json["installments_amount"]) ? $json["installments_amount"] : null;
+        $this->installmentsNumber = isset($json["installments_number"]) ? $json["installments_number"] : null;
+        $this->balance = isset($json["balance"]) ? $json["balance"] : null;
     }
 
     /**
@@ -70,6 +51,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $vci
+     *
      * @return TransactionStatusResponse
      */
     public function setVci($vci)
@@ -88,6 +70,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $amount
+     *
      * @return TransactionStatusResponse
      */
     public function setAmount($amount)
@@ -106,6 +89,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $status
+     *
      * @return TransactionStatusResponse
      */
     public function setStatus($status)
@@ -124,6 +108,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $buyOrder
+     *
      * @return TransactionStatusResponse
      */
     public function setBuyOrder($buyOrder)
@@ -142,6 +127,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $sessionId
+     *
      * @return TransactionStatusResponse
      */
     public function setSessionId($sessionId)
@@ -153,18 +139,19 @@ class TransactionStatusResponse
     /**
      * @return mixed
      */
-    public function getCardDetail()
+    public function getCardNumber()
     {
-        return $this->cardDetail;
+        return $this->cardNumber;
     }
 
     /**
-     * @param mixed $cardDetail
+     * @param mixed $cardNumber
+     *
      * @return TransactionStatusResponse
      */
-    public function setCardDetail($cardDetail)
+    public function setCardNumber($cardNumber)
     {
-        $this->cardDetail = $cardDetail;
+        $this->cardNumber = $cardNumber;
         return $this;
     }
 
@@ -178,6 +165,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $accountingDate
+     *
      * @return TransactionStatusResponse
      */
     public function setAccountingDate($accountingDate)
@@ -196,6 +184,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $transactionDate
+     *
      * @return TransactionStatusResponse
      */
     public function setTransactionDate($transactionDate)
@@ -214,6 +203,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $authorizationCode
+     *
      * @return TransactionStatusResponse
      */
     public function setAuthorizationCode($authorizationCode)
@@ -232,6 +222,7 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $paymentTypeCode
+     *
      * @return TransactionStatusResponse
      */
     public function setPaymentTypeCode($paymentTypeCode)
@@ -250,29 +241,12 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $responseCode
+     *
      * @return TransactionStatusResponse
      */
     public function setResponseCode($responseCode)
     {
         $this->responseCode = $responseCode;
-        return $this;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getInstallmentsNumber()
-    {
-        return $this->installmentsNumber;
-    }
-
-    /**
-     * @param mixed $installmentsNumber
-     * @return TransactionStatusResponse
-     */
-    public function setInstallmentsNumber($installmentsNumber)
-    {
-        $this->installmentsNumber = $installmentsNumber;
         return $this;
     }
 
@@ -286,11 +260,50 @@ class TransactionStatusResponse
 
     /**
      * @param mixed $installmentsAmount
+     *
      * @return TransactionStatusResponse
      */
     public function setInstallmentsAmount($installmentsAmount)
     {
         $this->installmentsAmount = $installmentsAmount;
+        return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getInstallmentsNumber()
+    {
+        return $this->installmentsNumber;
+    }
+
+    /**
+     * @param mixed $installmentsNumber
+     *
+     * @return TransactionStatusResponse
+     */
+    public function setInstallmentsNumber($installmentsNumber)
+    {
+        $this->installmentsNumber = $installmentsNumber;
+        return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getBalance()
+    {
+        return $this->balance;
+    }
+
+    /**
+     * @param mixed $balance
+     *
+     * @return TransactionStatusResponse
+     */
+    public function setBalance($balance)
+    {
+        $this->balance = $balance;
         return $this;
     }
 }
