--- src/Webpay/WebpayPlus/TransactionCommitResponse.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Transbank\Webpay\WebpayPlus;
-
-class TransactionCommitResponse extends TransactionStatusResponse
-{
-
-}
